[add]データベースクラスを追加
